+++ configs/ci/examples/lv_conf_examples.defaults
@@ -64,7 +64,6 @@ LV_USE_CALENDAR   1
 LV_USE_CALENDAR_CHINESE 1
 LV_USE_CANVAS     1
 LV_USE_CHART      1
-LV_USE_POLAR_CHART      1
 LV_USE_CHECKBOX   1
 LV_USE_DROPDOWN   1
 LV_USE_IMAGE      1
@@ -77,6 +76,7 @@ LV_USE_LIST       1
 LV_USE_LOTTIE     0
 LV_USE_MENU       1
 LV_USE_MSGBOX     1
+LV_USE_POLAR_CHART      1
 LV_USE_ROLLER     1
 LV_USE_SCALE      1
 LV_USE_SLIDER     1
