+++ configs/ci/perf/lv_conf_perf64b.defaults
@@ -38,7 +38,6 @@ LV_USE_BUTTONMATRIX  1
 LV_USE_CALENDAR   1
 LV_USE_CANVAS     1
 LV_USE_CHART      1
-LV_USE_POLAR_CHART      1
 LV_USE_CHECKBOX   1
 LV_USE_DROPDOWN   1   
 LV_USE_IMAGE      1 
@@ -53,6 +52,7 @@ LV_USE_LINE       1
 LV_USE_LIST       1
 LV_USE_MENU       1
 LV_USE_MSGBOX     1
+LV_USE_POLAR_CHART      1
 LV_USE_ROLLER     1  
 LV_USE_SCALE      1
 LV_USE_SLIDER     1
