+++ examples/widgets/polarchart/index.rst
@@ -1,0 +1,6 @@
+
+Line Polar Chart
+----------
+
+.. lv_example:: widgets/polarchart/lv_example_polar_chart_1
+  :language: c
