+++ lvgl.h
@@ -65,7 +65,6 @@ extern "C" {
 #include "src/widgets/calendar/lv_calendar.h"
 #include "src/widgets/canvas/lv_canvas.h"
 #include "src/widgets/chart/lv_chart.h"
-#include "src/widgets/chart/lv_polar_chart.h"
 #include "src/widgets/checkbox/lv_checkbox.h"
 #include "src/widgets/dropdown/lv_dropdown.h"
 #include "src/widgets/image/lv_image.h"
@@ -78,6 +77,7 @@ extern "C" {
 #include "src/widgets/lottie/lv_lottie.h"
 #include "src/widgets/menu/lv_menu.h"
 #include "src/widgets/msgbox/lv_msgbox.h"
+#include "src/widgets/polarchart/lv_polar_chart.h"
 #include "src/widgets/roller/lv_roller.h"
 #include "src/widgets/scale/lv_scale.h"
 #include "src/widgets/slider/lv_slider.h"
