+++ src/lv_conf_internal.h
@@ -2274,18 +2274,6 @@
     #endif
 #endif
 
-#ifndef LV_USE_POLAR_CHART
-    #ifdef LV_KCONFIG_PRESENT
-        #ifdef CONFIG_LV_USE_POLAR_CHART
-            #define LV_USE_POLAR_CHART CONFIG_LV_USE_POLAR_CHART
-        #else
-            #define LV_USE_POLAR_CHART 0
-        #endif
-    #else
-        #define LV_USE_POLAR_CHART      1
-    #endif
-#endif
-
 #ifndef LV_USE_CHECKBOX
     #ifdef LV_KCONFIG_PRESENT
         #ifdef CONFIG_LV_USE_CHECKBOX
@@ -2454,6 +2442,18 @@
         #endif
     #else
         #define LV_USE_MSGBOX     1
+    #endif
+#endif
+
+#ifndef LV_USE_POLAR_CHART
+    #ifdef LV_KCONFIG_PRESENT
+        #ifdef CONFIG_LV_USE_POLAR_CHART
+            #define LV_USE_POLAR_CHART CONFIG_LV_USE_POLAR_CHART
+        #else
+            #define LV_USE_POLAR_CHART 0
+        #endif
+    #else
+        #define LV_USE_POLAR_CHART      1
     #endif
 #endif
